Getters, Setters and properties in user

+++ src/main/java/com/avanshogeschool/API/repository/CarListingRepository.java
@@ -3,5 +3,5 @@ package com.avanshogeschool.API.repository;
 import com.avanshogeschool.API.domain.CarListing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarListingRepository extends JpaRepository<CarListing, Integer> {
+public interface CarListingRepository extends JpaRepository<CarListing, Long> {
 }
